Funciona todo el proceso de compra

+++ tickets/precio.py
@@ -1,6 +1,6 @@
 from django.db import models
 class Precio(models.Model):
-    precio = models.DecimalField(max_digits = 6, decimal_places = 2)
+    precio = models.IntegerField()
     sector = models.ForeignKey('Sector')
     espectaculo = models.ForeignKey('Espectaculo')
